+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ## Upcoming version
 
 ### Added
+
+- [[#40]](https://github.com/rust-vmm/vm-allocator/pull/40): Added serde
+  support for IdAllocator and AddressAllocator.
+- [[#99]](https://github.com/rust-vmm/vm-allocator/pull/99): Added APIs to
+  get the base address and size of an AddressAllocator.
+
 ### Changed
 ### Fixed
 ### Removed